Update search browser tests and move pagination tests to own file
